added ui for search bar
