+++ frontend/index.html
@@ -8,7 +8,9 @@
 
     <style>
       @import url('https://fonts.googleapis.com/css2?family=Playwrite+NG+Modern:wght@100..400&display=swap');
+      @import url('https://fonts.googleapis.com/css2?family=Inter+Tight:ital,wght@0,100..900;1,100..900&display=swap');
     </style>
+
 
   </head>
   <body>
