--- frontend/src/HomePage/Topbar.jsx
@@ -1,17 +1,0 @@
-import './Topbar.css'
-
-function Topbar() {
-    return (
-        <div className='bar'>
-            <div className='logo'>
-            <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="#47894b" class="bi bi-bar-chart-line-fill" viewBox="0 0 16 16">
-                <path d="M11 2a1 1 0 0 1 1-1h2a1 1 0 0 1 1 1v12h.5a.5.5 0 0 1 0 1H.5a.5.5 0 0 1 0-1H1v-3a1 1 0 0 1 1-1h2a1 1 0 0 1 1 1v3h1V7a1 1 0 0 1 1-1h2a1 1 0 0 1 1 1v7h1z"/>
-            </svg>
-            <p className='topname'>EconMetrics</p>
-
-            </div>
-        </div>
-    )
-}
-
-export default Topbar;
